blog-pipeline/pyproject.toml
[project]
name = "pipeline"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "authlib>=1.6.6",
    "bs4>=0.0.2",
    "fastapi[standard]>=0.128.0",
    "flask>=3.1.2",
    "imagehash>=4.3.2",
    "numpy>=2.4.1",
    "openpyxl>=3.1.5",
    "pandas>=3.0.0",
    "pillow>=12.1.0",
    "psycopg2>=2.9.11",
    "requests>=2.32.5",
]
